Create a GreetingService
We now extract the greeting feature from our runner and create a new
`GreetingService` interface that we implement in
`SimpleGreetingService`.

We're marking `SimpleGreetingService` as a `@Service` so Spring
considers it as an application component.

Since this feature is now extracted in its own component, we can test it
independently in our application.