Add a ShortURLRepository
Now that we've added `spring-boot-starter-data-mongodb`, we can create a
new `ShortURL` class that we'd like to save in a mongodb database.

For that, Spring Data offers `Repository` interfaces to interact with
your mongodb datastore. We've created here a `ShortURLRepository`.

We've also added an in-memory mongodb database (we could of couse use a
real mongodb instance here).