Create ShortURLs and save them in the database
To implement the first part of our service, we need to create `ShortURL`
instances from the given URLs and save them in our database.

For that, we can implement the `shortenURL` method from our
`ShortURLService` component. In that class, we need to:
* generate a random shortCode and create a ShortURL with it
* ask Spring to give us the `ShortURLRepository` component
* use that repository to save our `ShortURL` in the database

At that point the Tests do not compile yet, we need to implement the
next method first.