Redirect existing ShortURLs
We now need to add a new method to our controller to handle requests
like "GET /AbC" and respond with an HTTP redirect to the original URL.

For that, we can use `@GetMapping` to handle get requests on that new
method and fetch the shortCode information from the request path with
`@PathVariable`.
Once done, we can use this information to fetch any existing `ShortURL`
from the database using the `ShortURLService` and return it as the
response body.

Note: we've also added a `@ResponseStatus` annotation on our exception -
this tells Spring MVC how to translate that exception at the web layer
and use a HTTP 404 response status if that exception is thrown.